move helper class namespace
