+++ LibreNMS/Data/Source/Bird2.php
@@ -21,7 +21,7 @@
  * @link       https://www.librenms.org
  */
 
-namespace LibreNMS\Util;
+namespace LibreNMS\Data\Source;
 
 class Bird2
 {
